Make panel tests less likely to timeout on Mac when the test app loses the foreground status
Occasionally some panel tests timeout on Mac test bots. It seems that the app loses the foreground status for some reason and the panel testing code times out on waiting for activation signal. The fix is to try to detect this and force to activate the app before waiting on the panel activation notification.

BUG=none
TEST=existing tests

Review URL: https://chromiumcodereview.appspot.com/19018003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211083 4ff67af0-8c30-449e-8e8b-ad334ec8d88c