Pull methods in FileApiWorker to free functions.
This CL pulls methods in FileApiWorker to free functions, with taking an
extra argument FileSystemInterface*. By this change, we can pass the instance
of FileSystemInterface's instance at last.
For our use case, the methods' arguments are passed on IO thread, but the
FileSystem lives on UI thread, so it is necessary to post tasks between
IO and UI threads.
Currently, when the arguments come on IO thread, we binds the unsafe pointer
on IO thread, post the task to UI thread, check if file system is alive or not
on UI thread, and if yes invoke the closure.
However, after the refactoring to remove ChromeOS code from webkit/..., we need
to take the FileSystem pointer from Profile via DriveIntegrationService on UI
thread (i.e., we cannot use (even unsafe) pointer on IO thread).
This CL is the preparation of the refactoring. The new functions take
FileSystemInterface* "at the end of arguments", so we can bind arguments only
on IO thread, and feed FileSystem on UI thread later, easily.

BUG=253837
TEST=Ran unit_tests and trybot

Review URL: https://chromiumcodereview.appspot.com/18536004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211099 4ff67af0-8c30-449e-8e8b-ad334ec8d88c