Fetch policy for already signed-in users on Android.
If a user is already signed-in at startup but doesn't have policy cached then
a new registration for cloud policy will be performed.

This handles a couple of cases:
- signed-in users that updated Chrome to M30
- previous registration at sign-in time failed
- account wasn't managed before but is managed now

To avoid re-registering at every startup, the last registration attempt time
is saved in the prefs, and a new attempt is made at least 1 or 3 days later,
depending on the connection type.

BUG=249375

Review URL: https://chromiumcodereview.appspot.com/18801003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211138 4ff67af0-8c30-449e-8e8b-ad334ec8d88c