Remove incorrect DHCECK from Invalidations
The DCHECK was intended to make sure that we've bothered to fetch an
access token by the time we call UpdateInvalidatorCredentials(). That
seemed like a reasonable assertion.  However, if we're calling
UpdateInvalidatorCredentials with the intention of clearing the token
(which can happen when we receive NOTIFICATION_TOKENS_CLEARED) the
DCHECK will fail.

BUG=259320

Review URL: https://chromiumcodereview.appspot.com/18565007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211224 4ff67af0-8c30-449e-8e8b-ad334ec8d88c