Fix issue with backspace and tab not working in rAc sign-in
Removes the hack of registering an accelerator to prevent default
handling of backspace.
Sets focus initially to the sign-in view when displaying it.
Backspace handling was incorrect since the web view never got the
focus.
There is still an issue that the webview wont take the focus when
clicked; that is mitigated by this change, but still worth
investigating separately.  I'll be filing separate bugs for related
issues.

BUG=247612,242584
TEST=Validate 247612

Review URL: https://chromiumcodereview.appspot.com/18899004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211246 4ff67af0-8c30-449e-8e8b-ad334ec8d88c