Fixing excess indentation of policy table titles
This bug crept in when I started adding new policy table sections to the 'main-section' - since there is a css rule to give every section some indentation. Did the simplest possible fix by giving the main section an indentation of 0. See before & after fix screenshots attached to bug.

Is it ok that the table descriptions are still indented in relation to the titles now?

BUG=254935

Review URL: https://chromiumcodereview.appspot.com/18063002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211354 4ff67af0-8c30-449e-8e8b-ad334ec8d88c