Fix the allHistory flag on chrome.history.onVisitRemoved event.
Calling chrome.history.deleteAll results in chrome.history.onVisitRemoved
events to be dispatched with the allHistory flag being false instead of true.
This fix changes the call to ExpireHistoryBetween so that begin_time and
end_time are null which results in the history backend using a special path
to delete all history and dispatch the right event.

BUG=258445

Review URL: https://chromiumcodereview.appspot.com/18755003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211376 4ff67af0-8c30-449e-8e8b-ad334ec8d88c