Colors in views::StyledLabel.
Add the possibility to specify a custom foreground color when defining the style of a text range in a views::StyledLabel. Also support auto color readability in views::StyledLabel, to automatically adjust the foreground colors when they are not readable over the background color.

Necessary to comply with the mock of the bookmark sync promo (crbug.com/244279).

TEST=No user visible change introduced by this CL. Custom colors will be used in the bookmark sync promo (crbug.com/244279).
BUG=244630

Review URL: https://chromiumcodereview.appspot.com/17756003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@211471 4ff67af0-8c30-449e-8e8b-ad334ec8d88c