Change BrowserThreadDelegate to run Init() async.
After this change Init() method will be run as the first task
on the thread asynchronously from the UI thread initialization.
This allows us to speed up the startup by making IOThread Init()
a non-blocking task for the UI.

Changes AwUrlRequestContextGetter to create a CookieMonster on
UI thread since it needs to happen synchonously.

BUG=258231

Review URL: https://chromiumcodereview.appspot.com/18618004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212003 4ff67af0-8c30-449e-8e8b-ad334ec8d88c