Ask for user confirmation before signing in to a managed account.
If the account that the user is signing in to has management enabled then
a dialog is displayed that allows the user to go back and select another
account, or proceed to sign-in.

An Observer of the sign-in process is also introduced, so that the caller of
startSignIn() can react appropriately to cancelled signins.

This code is currently disabled behind a flag, and the strings are not final.

BUG=259354

Review URL: https://chromiumcodereview.appspot.com/19038004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212034 4ff67af0-8c30-449e-8e8b-ad334ec8d88c