Add CloudExternalDataStore
This CL adds a CloudExternalDataStore which will be used to persist
external data referenced by policies. The CloudExternalDataManager
class introduced in this CL is a stub that contains only a type definition
used by the CloudExternalDataStore. This class will be fleshed out in
follow-up CLs.

BUG=256635
TEST=New unit test

Review URL: https://chromiumcodereview.appspot.com/19267004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212062 4ff67af0-8c30-449e-8e8b-ad334ec8d88c