Cleaning up TODO(joi) entries in //net.
The two TODOs related to threading are apparently gone so I'm removing
the TODOs and associated code.

I no longer think it's worth pursuing the idea of separating
throttling and the sliding window concept, so transforming a TODO
related to that to just an observation.

TBR=darin@chromium.org
BUG=none

Review URL: https://chromiumcodereview.appspot.com/18261003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212129 4ff67af0-8c30-449e-8e8b-ad334ec8d88c