[sync] Clean up SyncSetupHandler: Detach sign in from sync setup
SyncSetupHandler is a class used for sync setup, and sometimes needs to
initiate a gaia sign in. Today, it does so via SigninTracker, and keeps
track of the sign in progress while displaying a spinner or a grayed out
"in progress" sign in button. This is unnecessary, since other classes
that initiate sync setup respond only after sign in is complete, instead
of keeping track of an ongoing sign in.

This patch does multiple things:
1) Removes the use of SigninTracker from SyncSetupHandler, and updates
   the logic to close the sync setup wizard when control is handed off
   to the gaia sign in tab.
2) Updates OneClickSigninHelper to automatically close the gaia sign in
   tab before handing back control to SyncSetupHandler, in cases when
   the settings page initiates a gaia re-auth.
3) Miscellaneous clean up related to the above behavior changes.

Note: This bug takes care of item #2 in bug 241353.

BUG=241353, 259150, 261317, 261367. 92948
TEST=Initiate sync setup or re-auth via the settings page. The setup dialog must be closed during sign in, and be reopened after sign in.

Review URL: https://chromiumcodereview.appspot.com/18698004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212279 4ff67af0-8c30-449e-8e8b-ad334ec8d88c