Remove ExtensionURLInfo, make security decisions in render process
When asking if an extension should have access to a given frame, we need to
consider the frame's URL and also if the frame is sandboxed. We check the latter
by asking if the frame's security origin is the unique origin. However, we can
only usefully do this in the render process when examining a frame. In the
browser process or other common code, there's no useful origin to use other than
one that duplicates information in the URL.

This does security checks in the render process before doing any URL-based
lookups and then uses URLs from that point on.

R=abarth, mpcomplete
BUG=259982,237267

Review URL: https://chromiumcodereview.appspot.com/16625012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212302 4ff67af0-8c30-449e-8e8b-ad334ec8d88c