Extension activity log database refactoring (step 2)
This is a continuation of work started in
https://codereview.chromium.org/18660004/ (committed as r211234).

Start work on merging all the activity types into a single database
table.  This patch creates just a single table and has each of the
Action types record to that table, however the class hierarchy of
multiple action types is still in place.  Eventually there will be just
a single Action type, but the merging of that code will happen in a
following patch.

As a stopgap, until the merger is done, there is a WatchdogAction
subclass which is used to represent records read back from the database.

BUG=255730

Review URL: https://chromiumcodereview.appspot.com/19234003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212429 4ff67af0-8c30-449e-8e8b-ad334ec8d88c