Make the rAc dialog stay for Autocheckout when it clicks on proceed button on the last page.
When there is a proceed element on the last page of Autocheckout flow, it will load the confirmation page which we do not have filemap for, so |TabAutofillManagerDelegate::DidNavigateMainFrame| will close the dialog. This CL added |AutofillManagerDelegate::InAutochekoutFlow()|, which will be called by |AutocheckoutManager| to explicitly notify the delegate if we are currently in an Autocheckout flow, and it should not close the dialog if the flow is not ended yet.

BUG=260062

Review URL: https://chromiumcodereview.appspot.com/19304003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212529 4ff67af0-8c30-449e-8e8b-ad334ec8d88c