Move the responsibility for enabling the app list to AppListServiceImpl.
Setting profiles is currently the job of AppListServiceImpl, but
responsibility for setting the profile _path_ is currently up to the the
platform-specific implementation. Similarly, enabling via the webstore
is accomplished in a platform-specific way.

After this CL, enabling and disabling the app list, and setting the app
list profile path, will be the responsibility of  AppListServiceImpl. To
do this, prefs::kAppLauncherHasBeenEnabled  moves from Windows-only to a
shared preference between "desktop" app  list implementations. 

BUG=138633, 261363
TEST=No functional changes

Review URL: https://chromiumcodereview.appspot.com/19091003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212570 4ff67af0-8c30-449e-8e8b-ad334ec8d88c