sync: Track tab node IDs in SyncedSessionTracker.
In order to build session sync atop Sync API, we'll need to track tab node
ids for tabs foreign sessions, primarily so we know which nodes to delete when
a foreign session becomes stale.  This patch adds tab_node_id to
SessionTabWrapper for this purpose.

In addition, some minor cleanup such as using a named enum parameter to the
Wrapper classes to denote ownership intent more clearly, and rewords some
comments and variables in SessionModelAssociator / SyncedSessionTracker to
improve readability and make it easier (hopefully :) ) to port to sync API.

BUG=98892, 80194
R=shashishekhar@chromium.org, zea@chromium.org

Review URL: https://codereview.chromium.org/19196003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212635 4ff67af0-8c30-449e-8e8b-ad334ec8d88c