Allow OSX App Launcher to be enabled via the webstore.
This implements IsAppLauncherEnabled for OSX, so the webstore can query
it, and removes the empty override for EnableAppList on OSX, so that
enabling it will do something.

The old, mac-only flag --enable-app-list-shim is still supported, but
only checked if the app list has not been enabled via the webstore.

BUG=138633, 261363
TEST=No functional changes on Win/CrOS. On Mac, running the initial
Chrome process with --enable-app-list will create (or recreate) an app
list shortcut, popup a finder window. Installing something from the
webstore will then popup the app launcher instead of highlighting the
app on the NTP when installed.

Review URL: https://chromiumcodereview.appspot.com/18736003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@212861 4ff67af0-8c30-449e-8e8b-ad334ec8d88c