Introduce EventRouter::DispatchEventWithLazyListener().
There's a pattern in the code where a lazy listener will be added to an
extension before dispatching things like onLaunched or onInstalled so
they can be dispatched to extensions that haven't been loaded yet. This
pattern is currently slightly wrong because it always results in there
being no lazy listener for the event, regardless of whether one existed
before (as listening on a lazy event is a binary proposition).

This change introduces DispatchEventWithLazyListener() to address this
problem and prevent the code duplication at the call sites.

BUG=224736

Review URL: https://chromiumcodereview.appspot.com/19752012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213045 4ff67af0-8c30-449e-8e8b-ad334ec8d88c