Reland "Move firefox_proxy_settings* out of importer."
Apparently FirefoxProxySettings has nothing to do with importer/, it's use only
by c/b/net. So per request we move it there.

This relands the commit at revision r212579.
It was reverted because it broke Android build given that it before it was being
excluded but after not. This fixes this issue, excluding the build of those
files from Android build again.

BUG=258876
TEST=unit_tests
R=gab@chromium.org,eroman@chromium.org,yfriedman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/19905003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213132 4ff67af0-8c30-449e-8e8b-ad334ec8d88c