Enable and disable Synced Notification Client Services
Synced Notifications can be sent from web services.  This checkin will allow us to enable and disable synced notifications from specific service sources.  For example, sending services could include mail, calendar, and Google+.  These notifications will be displayed to the user as Rich Notifications.

We are developing a protocol to send the names, app ids, and icons for these services as a new synced data type.  However, until that protocol is ready, we are hardcoding the names and icons of the services into chrome.  We will remove the hardcoded names and icons once the new sync datatype is ready (hopefully during the next milestone).

This change also supplies the name of our first client (Google+) and an icon to use in the synced notifications settings dialog.  Due to licensing restrictions, we are checking in one icon for chromium (basically a copy of the default favicon), and another one for chrome proper (this will be a separate checkin to the chrome specific codebase).

BUG=254617

Review URL: https://chromiumcodereview.appspot.com/19056002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213175 4ff67af0-8c30-449e-8e8b-ad334ec8d88c