[sync] Close duplicate overlay if user opens settings/syncSetup twice
If the user brings up the advanced sync settings page in a tab, and then
attempts to bring it up in another tab by directly navigating to
chrome://settings/syncSetup, we are supposed to transfer focus to the
first instance of the dialog because it is possible to register only one
instance of SyncSetupHandler as the handler of messages from the UI
layer.

However, before transfering focus to the original sync setup overlay, we
neglect to close the blank sync setup overlay in the second tab.

This patch adds logic to HandleShowSyncSetupUI to check if there is
already an existing sync setup dialog in a different tab, and if so, it
clsoes the blank overlay in the duplicate tab before transfering focus
to the original sync setup dialog.

BUG=261566
TEST=Open sync settings, and then navigate to settings/syncSetup in
another tab. Focus should be transferred back to the first tab, and
overlays should be closed on the second tab.

Review URL: https://chromiumcodereview.appspot.com/19846008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213179 4ff67af0-8c30-449e-8e8b-ad334ec8d88c