Adapting the UI to bring it closer to the spec, and fixing image fetching.
Firstly, this uses more correct fields for displaying synced notifications.
The Annotation field is better than the description or text fields in most
cases.

Secondly, this fixes some image fetch problems.  I was using images from the
expanded_notification section instead of the collapsed_notification section,
and the schemaless URLs were being interpreted by the chrome code as UNC paths!
I added a function to add a default schema of https: if none was provided to
prevent the GURL code from thinking "//" meant a UNC path on windows.

Thirdly, I noticed some changes to the server side protobuf that we were not in
sync with, so I copied the new fields into the protobuf that we use.  (This
includes the annotation field).

BUG=262303,247668

Review URL: https://chromiumcodereview.appspot.com/19771013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213288 4ff67af0-8c30-449e-8e8b-ad334ec8d88c