Use URL hashes to switch tabs of chrome://translate-internals/.
This CL enable to jump a tab directly by URL hash in chrome://translate-internals/.  For example, chrome://translate-internals/#detection-logs jumps to the detection-logs tab.  Likewise, when a user switch a tab, the URL's hash in the Omnibox will be changed to fit the current tab.

BUG=262870
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/19724006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213375 4ff67af0-8c30-449e-8e8b-ad334ec8d88c