FileAPI: Add Initialize() function to FileSystemBackend
This is a preliminary change for adding SyncFileSystemBackend.

At first, we planed to do initializing process for SyncFileSystemBackend in
InitializeFileSystem(), but it revealed that there are some hassles on that
way (see [*]), so we decieded to introduce a separate initialze function and
to rename InitializeFileSystem() to its original name (i.e. OpenFileSystem()).

[*] https://codereview.chromium.org/18668003/


BUG=242422
TEST=should pass all existing tests
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/19632004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213382 4ff67af0-8c30-449e-8e8b-ad334ec8d88c