Initial plumbing for Scroll End Effect
This CL adds a flag, scroll-end-effect, for controlling whether or not the
experimental scroll end effect is enabled. This flags is false by default. This
CL also adds in the needed plumbing to detect vertical overscroll occurring and
send information about it to BrowserView.
    
BUG=151356
TEST=Built image with squish effect patch and confirmed:
     - Feature turned off by default
     - Enabling the flag turns on the feature
     - Vertical overscroll causes the expected squish effect
     - Other overscroll effects are not broken

Review URL: https://chromiumcodereview.appspot.com/18265004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213534 4ff67af0-8c30-449e-8e8b-ad334ec8d88c