[sync] Fix crash in chrome://settings when sync is disabled
When sync is disabled via the --disable-sync flag, there are a couple of
spots in browser_options_handler.cc (r195410) and sync_ui_util.cc (r210618)
that neglect to check if the sync service is non-null, and therefore, cause
access violations.

This patch fixes those spots to either check for a non-null
ProfileSyncService object before accessing it, or to use the
ProfileKeyedService mechanism to access other classes like Profile
instead of using a null ProfileSyncService object.

BUG=259765
TEST=Disable sync, go to settings, and sign in to chrome. No resulting crash.

Review URL: https://chromiumcodereview.appspot.com/19566003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213783 4ff67af0-8c30-449e-8e8b-ad334ec8d88c