Rename ShortcutCreationPolicy to ShortcutCreationReason.
This is a more appropriate name because it describes what caused the
shortcut creation and allows each implementation to decide the best
behavior. E.g. duplication of shortcuts is only relevant to Windows.

This can also be easily extended to other cases, or to be more
granular, if necessary.

There should be no behavioral changes.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/19587008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213787 4ff67af0-8c30-449e-8e8b-ad334ec8d88c