Check the user_data_dir when deleting or launching a shim.
The CrAppModeUserDataDir entry in the Info.plist of app shim bundles
currently contain the app_data_path, i.e.
<user_data_dir>/<profile_dir>/Web Applications/_crx_<extension_id>/

This checks that the path starts with the user_data_dir, so that we don't delete
or launch shims made for a different user_data_dir.

BUG=168080
TEST=Install a packaged app.
Start Chrome with a different user_data_dir.
Install the same app.
Restart Chrome with the original user_data_dir.
Delete the app.
The shim for the app in /Applications/Chrome Apps/ should not be deleted.

Review URL: https://chromiumcodereview.appspot.com/19522003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213814 4ff67af0-8c30-449e-8e8b-ad334ec8d88c