Properly set showApps value in NTP.
Managed users do not currently have apps so they don't get app cards in the
new tab page. Currently a variable is set for managed users that indicates that
apps can be shown which triggers an assert in javascript NTP code.

This CL fixes that by setting that variable to false for MUs. It isn't set in
the shouldShowApps function as it is static and being managed is a
property of the profile.

BUG=249692
R=pam@chromium.org

Review URL: https://chromiumcodereview.appspot.com/19868004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213958 4ff67af0-8c30-449e-8e8b-ad334ec8d88c