Add tons of tests for the option page and fix overlay back/forward.
The behavior of the options page and its overlays was almost entirely
untested. Add a bunch of browser tests covering general behavior and
especially recent regressions (bugs 256997 and 248051).

Rename the OptionsUIBrowserTest and the OptionsBrowserTest to match
their respective filenames.

Options-page back/forward handling was relying on the behavior fixed in
r206418, in which parent pages were re-shown even if they were already
visible. That re-showing was hiding the topmost overlay as a side effect;
now that it's fixed, the overlay needs to be hidden manually (but without
updating the tab history).

BUG=256997
TEST=browser_tests --gtest_filter=OptionsWebUINavigationTest.*

Review URL: https://chromiumcodereview.appspot.com/19601002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213965 4ff67af0-8c30-449e-8e8b-ad334ec8d88c