Extension activity log database refactoring (step 3)
This is a continuation of work in
https://codereview.chromium.org/18660004/ and
https://codereview.chromium.org/19234003/.

Flatten the Extension::Action hierarchy, getting rid of all subclasses
and keeping just the top class.  The new Action class has a number of
data fields (some optional), so any of the old extension actions can be
encoded into it.  It maps in a straightforward fashion onto the current
database schema.

Change all the old sites that used specific Action subclasses to use the
new version, and fix up unit tests.

The extension activity log private API will be updated soon (by felt) to
more closely match the new Action class.  There is some basic
compatibility in place so that the end-to-end tests still pass, but the
private API unit tests have been disabled pending the rewrite.

BUG=255730
TBR=jhawkins@chromium.org

Review URL: https://chromiumcodereview.appspot.com/19690003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@213995 4ff67af0-8c30-449e-8e8b-ad334ec8d88c