[sync] Remove calls to OptionsPage.closeOverlay from C++
OptionsPage.closeOverlay is a JavaScript handler that is invoked when an
overlay dialog on a chrome://settings sub-page is closed. It has a side
effect of navigating from the sub-page back to the main
chrome://settings page.

As of today, SyncSetupHandler::OnDidClosePage invokes
OptionsPage.closeOverlay via a C++ -> JS call as part of closing the
overlay. While this is fine when the overlay is closed, it is not fine
if the user tries to navigate away from a settings sub-page to another
web page like www.google.com, since it redirects the navigation to
chrome://setting instead.

This patch removes all calls to OptionsPage.closeOverlay from the C++
layer, and makes sure that UI flows are not affected. It also fixes the
unit tests for SyncSetupHandler.

BUG=263747
TEST=Go to chrome://settings/syncSetup, then navigate to google.com.

Review URL: https://chromiumcodereview.appspot.com/20196003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214061 4ff67af0-8c30-449e-8e8b-ad334ec8d88c