DevToolsAdbBridge: get rid of AdbAttachCommand
When trying to attach to a page use recently fetched list of pages
instead of fetching the device list again.

The 'inspect' link is now always visible for ADB pages on chrome://inspect.
When clicked repeatedly it brings the existing devtools window to front.

BUG=None
R=pfeldman@chromium.org

Review URL: https://codereview.chromium.org/21041002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214186 4ff67af0-8c30-449e-8e8b-ad334ec8d88c