aura: Use ThemeService instead of ThemeProvider in browser actions.
This reverts most of the changes to BrowserActionView and switches it to
the ThemeService system instead of the ThemeProvider. This should fix a
crash in general (where the button gets updated by the extension system
while not part of a views hierarchy), and fixes stale theme images in
linux_aura (when switching between gtk and chrome themes).

BUG=130810, 264495
R=sky@chromium.org

Review URL: https://codereview.chromium.org/20574005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214214 4ff67af0-8c30-449e-8e8b-ad334ec8d88c