Wipe Profile data on signout if the account was managed, on Android.
The user is warned when signing in to an enterprise account that its data is
managed and owned by the enterprise domain; on sign out all the managed data is
gone, including bookmarks, history, cookies, etc.

The last signed in username is also cleared from the PrefService so that the
next sign in doesn't trigger the account switch dialog (which is unnecessary,
since all the data was removed and won't be mixed among the 2 accounts).

BUG=265450
R=mkwst@chromium.org, yfriedman@chromium.org

Review URL: https://codereview.chromium.org/21071002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214246 4ff67af0-8c30-449e-8e8b-ad334ec8d88c