Drive: Do not treat type-unknown deleted entries as an error.
When entries shared to the current user by other user is unshared,
such a change is reported as a 'deleted' entry with all other
metadata (including is_file() or is_folder()) unknown. To reflect
such deletion locally, we must not filter them out as an error.

Note that this fixes the BUG only for --enable-drive-v2-api case.

BUG=260020
R=hashimoto@chromium.org

Review URL: https://codereview.chromium.org/20559003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214293 4ff67af0-8c30-449e-8e8b-ad334ec8d88c