Omnibox: Pass Current Page Classification to Providers
This will allow, for instance, search provider to be more aggressive if
the user is on a search results page.

This change is actually small.  It merely looks big because it
changes the constructor to AutocompleteInput and that's used in
many places.

I also cleaned up a few comments while I'm touching these files.

When revising tests, I generally set the page classification to
INVALID_SPEC except for browser_tests and interactive_tests.  Those
tests are typically written to see how the browser behaves on startup.
For those, it felt more correct to set the page classification input to
NEW_TAB_PAGE.

TODO:
* create bug to track follow-up work to make this new parameter settable
  via about:omnibox
* create bug to track follow-up work to make AutocompleteClassifier get
  and pass along the current URL for its attempt to classify the query

TBR=pkasting@chromium.org

BUG=263823,237284

Review URL: https://codereview.chromium.org/20030003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214386 4ff67af0-8c30-449e-8e8b-ad334ec8d88c