[rAC, OSX] Implement autofill for rAC dialog.
This implements autofilling data in the rAC dialog from Chromium's
autofill data, as soon as a user interacts with a field that can be
autofilled.

(This is a re-land of r213532, which got reverted due to test failures. It
merely removes a NOTREACHED in detailInputForType:, so just TBR'd)

TBR=isherman@chromium.org, rsesek@chromium.org
BUG=256864

Review URL: https://codereview.chromium.org/21139002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214407 4ff67af0-8c30-449e-8e8b-ad334ec8d88c