Settings: Home page URL is not shown if closing home page overlay by Enter key
This fixes a problem similar to crbug.com/258345.

HomePageOverlay assumes a 'change' event is dispatched when it is
closed, before committing preference values in
SettingsDialog.handleConfirm(). 'blur' event for a URL field, which
triggers a 'change' event, is not dispatched if:
  * there is no focused element when opening the overlay, and
  * we close the overlay by Enter key
because OptionsPage.setOverlayVisible_() hides itself asynchronously and
the last focus is not restored.
We should make sure setOverlayVisible_(false) removes focus immediately.

BUG=263734

Review URL: https://chromiumcodereview.appspot.com/20141006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214558 4ff67af0-8c30-449e-8e8b-ad334ec8d88c