Change dialog texts for folder upload to explicitly indicate it's for 'Uploading'
Currently we show the same dialog title and confirm button
('Select Folder' and 'OK' or 'Open') for folder upload,
but it should more clearly tell that it's for 'uploading',
so that the user can know s/he is exposing the folder
content to the website.

This patch makes following UI string changes to make it clear:

* "Select a folder to upload" for dialog title on ChromeOS
(original text is: "Select a folder to open")
* "Select Folder to Upload" for dialog title on other platforms
(original text is: "Select Folder")
* "Upload" for dialog confirm button (original text is: "Open")

The new string texts have gotten LGTM by UI team.

BUG=252888
TEST=manual
R=avi@chromium.org, jamesr@chromium.org, sky@chromium.org, yoshiki@chromium.org

Review URL: https://codereview.chromium.org/18627002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214588 4ff67af0-8c30-449e-8e8b-ad334ec8d88c