WebUI: Do not show focus ring for radio buttons and checkboxes updated by labels.
This is a follow-up change for the behavior change by Blink r153039.

Settings page, extension setting page, and feedback page contains some
radio buttons and checkboxes associated with label elements, and we don't
want to show focus rings for them if they are mouse-clicked.

Canceling focus like crrev.com/213584 for each of labels is unrealistic.
So, we allow to focus on radio buttons and checkboxes, but don't show
focus ring for them with FocusOutlineManager.

The purpose of OptionsPage.showOverlay_ change is to avoid to focus on a
link in an overlay if we have a focused element without focus ring in the
base document.

BUG=264144,264567,264668,265327,266159
R=arv@chromium.org

Review URL: https://codereview.chromium.org/20556003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214614 4ff67af0-8c30-449e-8e8b-ad334ec8d88c