Cmd+Q on an app window quits the app.
In applicationShouldTerminate, if the terminate was triggered by a keyboard
event, and the currently focused window is a ShellWindow, then quit the app
associated with that window and return NSTerminateCancel.
An accessor for ExtensionAppShimHandler is added to
BrowserProcessPlatFormPart to allow manipulating apps from anywhere.
This will also be used for the Quit menu option for app menus. 
Cmd+Q behavior on a browser window remains mostly unchanged. The
only difference is that Cmd+Q when there are no windows quits immediately,
even if confirm-to-quit is enabled.

Note that this does not affect termination due to OSX shutting down. Even if
the shutdown is triggered by a keyboard shortcut (e.g. Ctrl+Cmd+Eject), the
event type will be NSSystemDefined, so Chrome terminates as usual.

BUG=168080,249059,227472
TEST=Start an app.
Cmd+Q with one of the app's windows focused should quit the app.
Quit Chrome and start an app using its shim.
Cmd+Q on the app should quit the app and Chrome.
TEST=Start an app.
Ctrl+Cmd+Eject with the app window focused to shut down OSX.
Chrome should terminate cleanly.

Review URL: https://chromiumcodereview.appspot.com/17448007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214672 4ff67af0-8c30-449e-8e8b-ad334ec8d88c