Clean up PnaclTranslationCache interface
This CL augments PnaclTranslationCache's public interface into a state
where it can be used by the PnaclHost. In particular, it adds GetKey
which generates a cache key from a cache_info object, and it switches
to using net::DrainableIOBuffer to transfer nexe data, which both
simplifies the implementation and reduces the number of times the data
needs to be copied.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3372
TEST= unit_tests --gtest_filter=PnaclTranslationCache*

Review URL: https://chromiumcodereview.appspot.com/21075004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214755 4ff67af0-8c30-449e-8e8b-ad334ec8d88c