Fixes bug where child Widgets wouldn't get deactivated on aura (2)
This is the second attempt at this. Here's the description:

This makes it so a native activation change is propagated to the
ActivationClient. This way child widgets see the activation change
(assuming they were active) and can update state. This means a child
bubble closes properly.

I'm holding off on the test for now as it requires an interactive ui
test.

Only difference between this and first attempt is I'm disabling a
test. For reasons I can't figure out the failing test fails in posting
a task. That is, we see the mouse event (from the hook we installed)
and post a task that'll make the test continue on. But the task never
gets processed (AFAICT it is never canceled nor is the object deleted
(which would also cancel the task)). I can't repro locally, which
makes it extremely difficult to figure out. So, I'm disabling now
since this doesn't seem related to my patch at all, maybe timing
related or something.

TBR since this is same as original.

BUG=262085
TEST=see bug
TBR=ben@chromium.org

Review URL: https://codereview.chromium.org/21354004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214849 4ff67af0-8c30-449e-8e8b-ad334ec8d88c