Add a flag to allow renderer to use software compositor when GL compositor doesn't work.
The --enable-software-compositing flag will allow the renderer to use the software compositor if it can't create a GL context. This is currently accomplished by having the compositor request a "fallback" output surface if the normal output surface can't be initialized (on the impl thread).


BUG=229712,230120

Review URL: https://chromiumcodereview.appspot.com/19267016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214930 4ff67af0-8c30-449e-8e8b-ad334ec8d88c