Update Mac notification tray behavior.
Now there are 8 icon resources instead of 2.  We will display different
icons based on 3 binary attributes:
* unread notifications
* do not disturb mode
* pressed status (message center open/closed)
The icon is now a bell (instead of a chat bubble), and we will no longer
display unread count with the notification tray icon.

Additionally, this patch causes the icon not to be updated while the
message center is open, and causes the icon not to be removed after the
first notification comes in (to make settings more accessible.)

BUG=264798,264971

Review URL: https://chromiumcodereview.appspot.com/21308002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214932 4ff67af0-8c30-449e-8e8b-ad334ec8d88c