Remove DirectoryContent.getPath().
We're going to switch the representation of a file/directory from URL/Path to
Entry to support multi-FileSystem.
This is the first step of the change. DirectoryContent.getPath() is equivalent
to DirectoryContent.getDirectoryEntry().fullPath, so this CL removes getPath()
by replacing the callee to getDirectoryEntry().fullPath.

BUG=265573
TEST=Ran browser_tests --gtest_filter="*FileSystemExtensionApiTest*:*FileManagerBrowserTest*"

Review URL: https://chromiumcodereview.appspot.com/21379003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@214940 4ff67af0-8c30-449e-8e8b-ad334ec8d88c