Add an in-memory container for registry trees.
This allows to pull registry trees into an in-memory representations, which
preserves the pecularities of the registry, in particular
case-insensitive-but-case-preserving semantics as well as potential name
clashes between key and value names. We'll use this as an intermediate format
when reading GPO from the registry or PReg files.

BUG=chromium:250860
TEST=Unit tests.

Review URL: https://chromiumcodereview.appspot.com/21363002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215021 4ff67af0-8c30-449e-8e8b-ad334ec8d88c