Stop registering PNaCl for updates on ChromeOS and reap old per-profile copies
Moving off of per-profile copies, and using the one copy in the system image.
The rest of the code will be deleted once we have this reaping enabled
for a while on M30 and hopefully on M29.

BUG=221381
TEST=manually create a profile w/ PNaCl (w/out patch), and then apply patch
and check that it gets deleted
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/20992007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215053 4ff67af0-8c30-449e-8e8b-ad334ec8d88c