Refresh access tokens for TiclInvalidationService
The Google Talk server will reject connection attempts made with a stale
access token.  This is often not a problem because we only need to
connect on startup, when the token is fresh.  It becomes a problem if we
temporarily lose the network connection and need to reconnect.

This change updates the TiclInvalidationService so it requests an
updated access token when it receives a credentials rejected error.
This should allow it to recover from connection failures caused by stale
access tokens.

TEST=Manual.  Sign in to sync, wait > 1 hour, disable network
connection, then re-enable it.  Verify notifier state is enabled.

BUG=264455

Review URL: https://chromiumcodereview.appspot.com/21140004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215094 4ff67af0-8c30-449e-8e8b-ad334ec8d88c