More UI fixes for Synced Notifications
The text in Synced Notifications still does not match the spec, this will
bring it closer.  The problem here is that the UI mocks for Synced
Notifications have a heading and two lines of text, but the Rich
Notifications framework only has one line of text.

The proposed solution is to concatenate the two text fields with a newline.

This checkin also removes some GUNS hardcoding that we had previously.

BUG=266966

Review URL: https://chromiumcodereview.appspot.com/21586002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215393 4ff67af0-8c30-449e-8e8b-ad334ec8d88c