sync: Remove SyncSourceInfo
The SyncSourceInfo was a struct that contained a GetUpdatesSource and a
ModelTypeInvalidationMap.  Both of these types are in the process of
being deprecated.  The SyncSourceInfo itself was used only for debugging
(about:sync), tests (mostly sync_scheduler_unittest.cc) and maintaining
compatibility with some old function signatures.

Removing the SyncSourceInfo allow us to remove dependencies on
ModelTypeInvalidationMap, which is a step towards enabling invalidation
"trickles" support.

BUG=233437

Review URL: https://chromiumcodereview.appspot.com/19982002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215446 4ff67af0-8c30-449e-8e8b-ad334ec8d88c