Omnibox: Create Bundled Field Trial; Convert SearchHistory trial to it
Create one field trial that's intended to rule over all omnibox field
trials.  They allows our trials to run correlated.  For example, suppose
we have two trials A (with 50-50 groups a1 and a2) and B (with 50-50
groups b1 b2).  Currently this yields four conditions users can be in
25% a1-b1, 25% a1-b2, 25% a2-b1, 25% a2-b2.

This would be great, but what if we want to compare a1-b1 and a1-b2 and
a2-b1 but think a2-b2 makes no sense.  There's no way to do it with the
current independent field trial framework.

This change creates one field trial that supposed to include all omnibox
field trials that we want to run.

This change converts the SearchHistory field trial to this bundled
field trial.

Obviously, this trials-can-be-dependent property isn't useful until we 
add more experiments to this bundled field trial.  I'll do that in 
another changelist.

This change also gives this new bundled trial the ability to condition 
certain experiment on page context.  In particular, depending on what
the user is viewing when using the omnibox, the different experiments
can be active.  For example, an omnibox on the new tab page can have
different experiments than the omnibox on an arbitrary web page.

Tested using my own variations server with an appropriate config
with various experiment params.

BUG=264062,264065,263823

Review URL: https://chromiumcodereview.appspot.com/20777006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215457 4ff67af0-8c30-449e-8e8b-ad334ec8d88c