Use an app shim in the user data directory for the app launcher.
This improves the cohesion between the kAppLauncherHasBeenEnabled
preference and the app shortcut that is added to the dock.

BUG=138633
TEST=From a fresh install, enable the app launcher on OSX. Right-click
the icon added to the dock, and select 'Show in Finder'. Chrome's user
data dir should be revealed, rather than a subfolder of the Profile
directory.

Review URL: https://chromiumcodereview.appspot.com/21990003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215603 4ff67af0-8c30-449e-8e8b-ad334ec8d88c