Prevent runaway progressbar on copying in Files.app
This patch fixed the bug described in crbug.com/267218.

The reason of the bug is that the pending (remaining) size is subtracted twice wrongly: the first is on progress event and the second is on completed event. It causes the value over-subtracted, and the pending bytes becomes negative value. This patch fixed this problem by reworking of the code of pending/completed sizes calculation.

BUG=267218
TEST=manual
R=hidehiko@chromium.org, mtomasz@chromium.org

Review URL: https://codereview.chromium.org/21756003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215620 4ff67af0-8c30-449e-8e8b-ad334ec8d88c