Fix OneClickSigninHelper::ShowInfoBarIfPossible Crash.
Fixes a crash that occurred by navigating to
https://accounts.google.com/?service=chromiumsync&continue= <invalid url>.

The problem
SyncPromoUI::GetSourceForSyncPromoURL passes the invalid URL to net::GetValueForKeyInQuery, which calls GURL::spec() on it. A DCHECK is made in GURL::spec() to check that the URL is valid. A crash occurs when it's not the case.

The fix
Make net::GetValueForKeyInQuery() return false instead of crashing when it receives an invalid URL. Similarly, many methods from GURL (query(), host(), port(), path(), ref()) don't crash when called on an invalid URL.

With this change, SyncPromoUI::GetSourceForSyncPromoURL will return SOURCE_UNKNOWN when it receives an invalid URL. This is the existing behavior for the case where the continue URL is valid, but unknown.

TEST=
1. Navigate to https://accounts.google.com/?service=chromiumsync&continue=http:%01%3F%01
Expected result: Chromium doesn't crash and an account management page
is loaded.
BUG=247977

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=215309

Review URL: https://chromiumcodereview.appspot.com/20491002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215627 4ff67af0-8c30-449e-8e8b-ad334ec8d88c