Turn off Synced Notifications at display level instead of network level.
To allow load testing of Synced Notifications as we ramp up, clients must
actually ask for the synced datatype even before we turn the feature on.

This change allows the network traffic of Synced Notifications even to machines
where the user has not yet turned on Synced Notifications.  The notifications
are discarded before displaying them, so there is no user-visible impact to
this change.  This change does mean that the Synced Notifications code will be
 exercised even when it has not been enabled specifically by the user.

The intent here is to allow this to be in the code for one milestone so that we
can slowly ramp up the white list before enabling the feature in the following
milestone.  When we enable the feature, we will remove the switch or undo this
change list.

Without this feature, we have no way to load test until the feature ships in the
wild, and we would just have to hope that we had fixed all the bugs and have the
capacity for launch.

BUG=267782

Review URL: https://chromiumcodereview.appspot.com/21950006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215754 4ff67af0-8c30-449e-8e8b-ad334ec8d88c