Introduce fileOperationUtil.
This CL introduces fileOperationUtil namespace and move/extracts some helper
methods from FileCopyManager to it.
This is preparation to move copyTo helper to it, so that moving the copyTo
from js-side to c++-side should be simpler task. The moving copyTo will be done
in following CLs.

BUG=246976
TEST=Ran browser_tests --gtest_filter="*FileSystemExtensionApiTest*:*FileManagerBrowserTest*" and tested manually.

Review URL: https://chromiumcodereview.appspot.com/22291004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215854 4ff67af0-8c30-449e-8e8b-ad334ec8d88c