Translate: browser test for Chrome Translate in an isolated world
An browser test which cover whole translate process was disabled tentatively.
This is because now Translate scripts runs inside a v8 isolated world and
it make impossible to run with local testing server.
This change introduce a flag to change the allowed server for XHR, and
make the test enabled again.
Note that this flag is warned by infobar when it is enabled.

BUG=268785
TEST=browser_tests --gtest_filter='Translate*.*'
TBR=hajimehoshi@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14180010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215912 4ff67af0-8c30-449e-8e8b-ad334ec8d88c