Change the meaning of SelectFileDialog.support_drive: it means Drive-aware callers.
Previously, what the |support_drive| flag means is just to show
Google Drive tab in the file dialog; it doesn't change other
behaviors.

Now, we are going to enable the "Google Drive" tab on Chrome OS
for all call sites of the dialog, by adding a feature to Drive client
to provide local snapshot so that callers need not care about Drive.

On the other hand, the "create snapshot" trick is unnecessary for
clients like Packaged app or extensions that implement fileBrowserHandler,
which access files through HTML5 FileSystem API.
So, for those clients I want to leave the |support_drive| flag to
indicate that no special handling is needed. Here's the summary:

Before:
- support_drive, folder-dialog : show Drive tab. No special handling.
- support_drive, open-dialog : show Drive tab. Create local snapshot.
- support_drive, save-dialog : show Drive tab. No special handling.
- !support_drive, folder-dialog : hide Drive tab.
- !support_drive, open-dialog : hide Drive tab. (this mode is not used)
- !support_drive, save-dialog : hide Drive tab.

After:
- support_drive, folder-dialog : show Drive tab. No special handling.
- support_drive, open-dialog : show Drive tab. No special handling.
- support_drive, save-dialog : show Drive tab. No special handling.
- !support_drive, folder-dialog : *HIDE* Drive tab. (this is not supported yet)
- !support_drive, open-dialog : Show Drive tab. Create local snapshot.
- !support_drive, save-dialog : *HIDE* Drive tab. (VERY SOON switch to => "Show Drive tab. Create local snapshot.")

BUG=140425

Review URL: https://chromiumcodereview.appspot.com/22165002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@215922 4ff67af0-8c30-449e-8e8b-ad334ec8d88c