Clean up compositor initialization/destruction.
Moved from https://codereview.chromium.org/21052007/

Currently the ContextFactory is outliving the compositor thread in
tests, which is bad since the contexts in there can be bound to the
compositor thread. Then we end up reusing those contexts and bad
things happen.

This enforces ordering for initializing and destroying the
compositor as follows:

Initialize ContextFactory
..Compositor::Initialize
....Create Compositor instances
....Delete Compositor instances
..Compositor::Terminate

The Compositor::Terminate call also destroys the ContextFactory.

The ContextFactory can be initialized in one of two ways:
1. ImageTransportFactory::Initialize will set up the
ContextFactory. This is used by things that invoke all of
content/browser/.
2. Compositor::InitializeContextFactoryForTests() must be
explicitly called by any unit tests that create a compositor.

Since some tests want a real GL context and some don't, this
does away with the misnomer of initializing the Compositor once
for the entire test suite, and then re-initializing somewhere
in the middle of the test suite. Instead, each test must
Initialize/Terminate the compositor with the ContextFactory
type of its choice.

Incidently, this test enables 20 tests on aura or win_aura that
were previously being skipped.

TBR=piman@chromium.org
BUG=258625, 266565

Review URL: https://codereview.chromium.org/22293003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216179 4ff67af0-8c30-449e-8e8b-ad334ec8d88c