Clone a subset of content::DevToolsProtocol class into DevToolsProtocol class in chrome
There is a growing number of places under chrome/browser/devtools where protocol messages
are handled in the browser process (as opposed to the renderer). Manual composition
and parsing of messages is making the code less readable and content::DevToolsProtocol
utility class is not accessible.

BUG=None
TBR=pfeldman@chromium.org

Review URL: https://codereview.chromium.org/22274018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216187 4ff67af0-8c30-449e-8e8b-ad334ec8d88c