Cleanup sync promo preferences (rename "Sync" to "SignIn" and remove unused preference).
It's possible to sign in to Chromium without enabling sync. For that reason, the name of the preferences related to sign in promos now start with kSigninPromo... instead of kSyncPromo...

Also, the kSyncPromoErrorMessage preference is no longer used so it has been removed from the "prefs" namespace. It's now cleared in browser_prefs::MigrateUserPrefs. When it's safe to do so, it will be completely removed.

TEST=No user visible change.
BUG=264283
TBR=gab@chromium.org,battre@chromium.org,sky@chromium.org,estade@chromium.org

Review URL: https://chromiumcodereview.appspot.com/22212003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216207 4ff67af0-8c30-449e-8e8b-ad334ec8d88c