Fix a logic error in an unittest.
This issue was found by a linter.


test_cases_ is always true (EXPECT_TRUE).

Even if it is false (and true), this statement cause a null deferencement...

   if (!test_cases_ && current_test_case_ >= test_cases_->size())

So, I believe the logic is to remove the negation. The test may be kept in case expect failed.
Otherwise, we should use ASSERT.



R=tbarzic@chromium.org, hashimoto@chromium.org

BUG=

Review URL: https://chromiumcodereview.appspot.com/22290008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c