Adds a SqlLite backend to WebRTCIdentityStore. WebRTCIdentityStore maintains a list of inflight requests. If a new request is identical (i.e. same origin, identity_name, common_name) to a inflight request, the new one is joined into the existing one, so that it will not make duplicated requests to the backend or generate dupicated identities.
If the backend does not find an existing identity, WebRTCIdentityStore generates a new one and adds it to the backend.

The backend borrows most of the design from SQLiteServerBoundCertStore::Backend. It loads the database into memory on the first find request. Add or delete database operations are batched to commit every 30 seconds or every 512 operations.


BUG=

Review URL: https://chromiumcodereview.appspot.com/21453002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216264 4ff67af0-8c30-449e-8e8b-ad334ec8d88c