Remove unused code in DirectoryTree in Files.app.
Historically, the directory tree on Files.app was used not only for Google Drive volume but also non-drive volumes. But now, it is only for the drive volume.

Major changes:
- Rename DirectoryTreeUtil.shouldHideTree -> isEligiblePathForDirectoryTree
- Hard-code the top level items in the tree.
- Make the tree itself not have its entry or its path.

This patch removes the unused code in the DirectoryTree class. Of cause, we can do more removal in directory_tree.js, but as for now, this removes some of them.

BUG=269903
TEST=The directory tree still works.
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/22631004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216385 4ff67af0-8c30-449e-8e8b-ad334ec8d88c